cast to in merge segmentations

+++ spatial_data/pp/utils.py
@@ -258,7 +258,7 @@ def merge_segmentation(s1, s2, label1=1, label2=2, threshold=1.0):
     # make final mask
     clean_mask = _remove_unlabeled_cells(s3, selected_cells)
     final_mask, fmap, bmap = relabel_sequential(clean_mask)
-    mapping = dict(zip([fmap[i] for i in selected_cells], [label1] * len(i1) + [label2] * len(i2)))
+    mapping = dict(zip([fmap[i] for i in selected_cells.astype(int)], [label1] * len(i1) + [label2] * len(i2)))
 
     return final_mask, mapping
